move ClassLiteral and FieldSelect from core/lombok/core/AnnotationValues.java to utils/lombok/core/
