+++ src/utils/lombok/core/ClassLiteral.java
@@ -1,0 +1,13 @@
+package lombok.core;
+
+public class ClassLiteral {
+    private final String className;
+
+    public ClassLiteral(String className) {
+        this.className = className;
+    }
+
+    public String getClassName() {
+        return className;
+    }
+}
