+++ src/utils/lombok/core/FieldSelect.java
@@ -1,0 +1,13 @@
+package lombok.core;
+
+public class FieldSelect {
+    private final String finalPart;
+
+    public FieldSelect(String finalPart) {
+        this.finalPart = finalPart;
+    }
+
+    public String getFinalPart() {
+        return finalPart;
+    }
+}
